requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
cloudinary==1.44.1
coverage==7.10.7
cryptography==46.0.5
dj-database-url==3.0.1
Django==4.2.28
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==37.12.0
gunicorn==23.0.0
iniconfig==2.1.0
MarkupSafe==3.0.3
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.11.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-http-client==3.3.7
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
whitenoise==6.11.0
